+++ cli/aws_orbit/models/manifest.py
@@ -58,7 +58,6 @@ class TeamManifest:
     jupyterhub_inbound_ranges: List[str] = field(default_factory=lambda: ["0.0.0.0/0"])
     image: Optional[str] = None
     plugins: List[PluginManifest] = field(default_factory=list)
-    profiles: List[Dict[str, Union[str, Dict[str, Any]]]] = field(default_factory=list)
     efs_life_cycle: Optional[str] = None
     authentication_groups: Optional[List[str]] = None
 
@@ -193,6 +192,7 @@ class NetworkingManifest:
 class FoundationManifest:
     Schema: ClassVar[Type[Schema]] = Schema
     name: str
+    role_prefix: Optional[str] = None
     codeartifact_domain: Optional[str] = None
     codeartifact_repository: Optional[str] = None
     images: FoundationImagesManifest = FoundationImagesManifest()
@@ -205,6 +205,7 @@ class FoundationManifest:
 class Manifest:
     Schema: ClassVar[Type[Schema]] = Schema
     name: str
+    role_prefix: Optional[str] = None
     user_pool_id: Optional[str] = None
     scratch_bucket_arn: Optional[str] = None
     eks_system_masters_roles: Optional[List[str]] = cast(List[str], field(default_factory=list))