+++ samples/notebooks/M-Admin/extra_user_apps/apps/lens/start_lens.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+sudo chmod +x /opt/orbit/codeserver_proxy/create_kubeconfig.sh
+source /opt/orbit/codeserver_proxy/create_kubeconfig.sh
+
+sudo chmod a+x /opt/orbit/extra_use_apps/apps/lens/Lens-5.1.3.AppImage
+./Lens-5.1.3.AppImage --appimage-extract-and-run